Reorganize phase6 e2e tests into domain-specific test files
Fold phase6-journey.spec.ts tests into itinerary-journey and trip-journey
so test file names reflect what they test, not which dev phase they came from.
